feat: split up the api service
